Move RadioButton to separate file and add tests
